requirements.txt
openai==1.3.5
wikipedia==1.4.0
python-dotenv==1.0.0